feat!: Add version of Graph that handles fully-parsed commits
This renames `graph::Commit` to `graph::LazyCommit` to make space for `graph::Commit` to be a fully owned.
`LazyCommit::to_owned()` was added to obtain fully owned `Commit` instances.
Rename `Graph::try_lookup_and_insert()` to `Graph::try_lookup_or_insert()` and
`Graph::try_lookup_and_insert_default()` to `Graph::try_lookup_or_insert_default()`

Additionally, add the `peek()` and `iter_unordered()` method to the `PriorityQueue`, along with an implementation for `Clone`
Rename `PriorityQueue::iter_random()` to `::iter_unordered()`.
