+++ gix-revision/tests/revision.rs
@@ -1,4 +1,5 @@
 mod describe;
+mod graph;
 mod spec;
 pub type Result<T = ()> = std::result::Result<T, Box<dyn std::error::Error + 'static>>;
